% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amgfl_CI.R
\name{amgfl.CI}
\alias{amgfl.CI}
\title{Simultaneous confidence intervals via the bootstrap}
\usage{
amgfl.CI(
  res,
  N = 100,
  Ndiv = 100,
  .alpha = 0.05,
  seed = NULL,
  progress = FALSE
)
}
\arguments{
\item{res}{the output of \code{amgfl}}

\item{N}{the number of bootstrap sampling}

\item{Ndiv}{the number of divisions for intervals}

\item{.alpha}{calculate (1-\code{.alpha})-simultaneous confidence intervals}

\item{seed}{the first parameter of \code{set.seed}}

\item{progress}{if TRUE, the progress of bootstrap is displayed}
}
\value{
a list having the following elements:
\item{trends}{a list having figures of trends with simultaneous confidence
for each variable of \code{X2}}

\item{trends.all}{a figure consists of all figures of \code{trends}}

\item{cluster}{a figure consists of three panels;
the left expresses the lower of simultaneous confidence intervals;
the middle expresses \code{mu} and the estimated cluster;
the right expresses the upper of simultaneous confidence intervals}
}
\description{
\code{amgfl.CI} This function provides simultaneous confidence intervals
via the bootstrap
}
\examples{
#amgfl.CI(res)
}
