% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amgfl_pred.R
\name{amgfl.pred}
\alias{amgfl.pred}
\title{Estimation for future observations}
\usage{
amgfl.pred(res, X1 = NULL, X2, area = NULL, Sp = NULL)
}
\arguments{
\item{res}{the output of \code{amgfl}}

\item{X1}{future observation: a matrix of explanatory variables, without an intercept, and with colnames,
where the variables are fitted with a linear model;
the intercept is automatically added}

\item{X2}{future observation: a matrix of explanatory variables with colnames,
where the variables are fitted with an additive model}

\item{area}{future observation: a vector of integers expressing sub-areas for each individual;
if \code{Sp} is given, \code{area} can be omitted}

\item{Sp}{future observation: a matrix with two columns expressing coordinates of observed points;
if \code{area} is given, \code{Sp} can be omitted}
}
\value{
a vector of predictive values
}
\description{
\code{amgfl.pred} This function provides predictive values for future observations
}
\examples{
#amgfl.pred(res, X1, X2, area, Sp)
}
