% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amgfl.R
\name{amgfl}
\alias{amgfl}
\title{Trend estimations by semiparametric additive model
with graph trend filtering via generalized fused Lasso (v0.7.0)}
\usage{
amgfl(
  dataset,
  space,
  nknots = NULL,
  scale1 = FALSE,
  scale2 = FALSE,
  power1 = 2,
  power2 = NULL,
  maxit = 100,
  tol = 1e-08,
  progress = FALSE,
  complete = TRUE
)
}
\arguments{
\item{dataset}{a list \code{list(y=y, X1=X1, X2=X2)} of which \code{y} is a vector of
a response variable, \code{X1} is a matrix of explanatory variables, without
an intercept, and with colnames, where the variables are fitted with a
linear model, and \code{X2} is a matrix of explanatory variables with colnames,
where the variables are fitted with an additive model;
\code{X1} can be omitted and the intercept is automatically added into \code{X1}}

\item{space}{a list expressing spatial information;
if a spatial division is given, \code{space} is given by the form of
\code{list(area=area, adjacent=adjacent, Sp=Sp, div.pol=div.pol)},
where \code{area} is a vector of integers expressing sub-areas for each individual,
\code{adjacent} is a matrix with two columns expressing adjacent relationships among sub-areas,
\code{Sp} is an optional parameter which is a matrix with two columns expressing
coordinates of observed points, and \code{div.pol} is an optional parameter which
is an sfc_POLYGON object expressing a spatial division;
if not, i.e., a spatial division is given by \code{amgfl}, \code{space} is given by the form of
\code{list(Sp=Sp, m=m, spB=spB, subregion=subregion, seed=seed)};
where \code{Sp} is the same of above but must be required,
\code{m} is an optional parameter expressing the number of sub-areas,
\code{spB} is an optional parameter which is an sfc_POLYGON object expressing
a boundary of the space subjected to analysis, \code{subregion} is an optional
parameter which is a vector expressing subregions for each individual,
and \code{seed} is an optional parameter which is the first parameter of \code{set.seed};
if \code{subregion} is given, \code{m} is a vector of the numbers of sub-areas for
each subregion and \code{spB} is an sfc_POLYGON object expressing subregion-wise
boundary and must be given}

\item{nknots}{the number of knots}

\item{scale1}{if \code{TRUE}, \code{scale} is applied to \code{X1}}

\item{scale2}{if \code{TRUE}, \code{scale} is applied to \code{X2}}

\item{power1}{a positive value adjusting the strength of the penalty in EGCV for
selecting tuning parameter in generalized fused Lasso}

\item{power2}{a positive value adjusting the strength of the penalty in EGCV for
selecting the number of eigenvalues in nonparametric smoothing; default is \code{log(length(y))}}

\item{maxit}{an upper bound for iteration}

\item{tol}{a tolerance for convergence}

\item{progress}{if \code{TRUE}, estimation progress is displayed}

\item{complete}{if \code{TRUE}, the message is displayed when the process terminates}
}
\value{
a list having the following elements:
\item{mu}{a vector of the estimates of spatial effects}

\item{beta}{a vector of the estimates of coefficients for \code{X}}

\item{alpha}{a vector of the estimates of coefficients for \code{B}}

\item{fitted.values}{a vector of the fitted values}

\item{OLS}{results of the ordinary least square (OLS) estimation given by the form of
a list \code{list(mu, beta, alpha, fitted.values)},
where \code{mu}, \code{beta}, and \code{alpha} are the OLS estimates
and \code{fitted.values} is a vector of fitted values based on the OLS estimates}

\item{summary}{a summary of estimation results given by the form of
a data.frame \code{data.frame(lambda, eigenvalues, division, group, runtime, iteration, convergence)},
where \code{lambda} is the optimal tuning parameter,
\code{eigenvalues} is the optimal number of eigenvalues,
\code{division} is the number of sub-areas,
\code{group} is the number of estimated clusters,
\code{runtime} is computational time (seconds),
\code{iteration} is the number of iterations until convergence,
and \code{convergence}: whether the algorithm converges}

\item{dataset}{a list of dataset \code{list(y, X1, X2)}}

\item{space}{a list expressing a spatial information}

\item{par}{a list of the entered parameters}

\item{scale}{a list of attributes of \code{X1} and \code{X2} given by applying \code{scale}}

\item{X}{a centralized matrix for which coefficients are estimated without the penalty}

\item{B}{a centralized matrix for which coefficients are estimated with the penalty}

\item{center}{a list of vectors using for centering of \code{X} and \code{B}}

\item{knots}{a matrix of knots}

\item{trends}{a list having figures of trends for each variable of \code{X2}}

\item{trends.all}{a figure consists of all figures of \code{trends}}

\item{division}{a figure expressing the spatial division}

\item{cluster}{a figure expressing \code{mu} and the estimated cluster}
}
\description{
\code{amgfl} This function provides trend estimation results
by semiparametric additive model with graph trend filtering via generalized fused Lasso
}
\examples{
#amgfl(dataset, space, nknots)
}
