% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exSR.R
\docType{data}
\name{exSR}
\alias{exSR}
\title{An example subregion}
\format{
An sf with 4 rows and 2 columns:
\describe{
\item{subregion}{subregions}
\item{geometry}{sfc_POLYGON}
}
}
\source{
none
}
\usage{
exSR
}
\description{
This is en example space expressing subregions in \code{exData}
}
\keyword{datasets}
